// src/main/webapp/ui/src/stores/models/Barcode.ts
export class PersistedBarcode implements BarcodeRecord {
  id: Id;
  data: BarcodeValue;
  description: string;
  newBarcodeRequest: boolean;
  deleteBarcodeRequest: boolean;
  imageUrl: ?URL;
  image: ?File;

  constructor(attrs: PersistedBarcodeAttrs) {
    makeObservable(this, {
      id: observable,
      data: observable,
      description: observable,
      imageUrl: observable,
      image: observable,
      markAsDeleted: action,
      setDescription: action,
      paramsForBackend: computed,
      isDeleted: computed,
    });
    if (attrs.newBarcodeRequest) {
      this.id = null;
      this.newBarcodeRequest = true;
      this.deleteBarcodeRequest = false;
      this.description = attrs.description;
    } else if (attrs.deleteBarcodeRequest) {
      this.id = attrs.id;
      this.newBarcodeRequest = false;
      this.deleteBarcodeRequest = true;
      this.description = attrs.description;
      this.imageUrl = attrs.imageUrl;
    } else {
      this.id = attrs.id;
      this.newBarcodeRequest = false;
      this.deleteBarcodeRequest = false;
      this.description = attrs.description;

      this.image = null;
      const imageUrl = attrs._links.find(
        ({ rel }) => rel === "enclosure"
      )?.link;
      this.imageUrl = imageUrl;
    }
    this.data = attrs.data;
  }

  fetchImage(): Promise<File> {
    if (!this.imageUrl) return Promise.reject();
    return fetchImage(this.imageUrl, this.description);
  }

  markAsDeleted(): void {
    this.deleteBarcodeRequest = true;
  }

  deletedCopy(): ?BarcodeRecord {
    if (this.id === null) return null;
    return new PersistedBarcode({
      id: this.id,
      deleteBarcodeRequest: true,
      data: this.data,
      description: this.description,
      imageUrl: this.imageUrl,
    });
  }

  get paramsForBackend(): { ... } {
    const params: {
      newBarcodeRequest?: true,
      deleteBarcodeRequest?: true,
      ...
    } = {
      id: this.id,
      data: this.data,
      description: this.description,
    };
    if (this.newBarcodeRequest) params.newBarcodeRequest = true;
    if (this.deleteBarcodeRequest) params.deleteBarcodeRequest = true;
    return params;
  }

  setDescription(value: string): void {
    this.description = value;
  }

  get isDeleted(): boolean {
    return this.deleteBarcodeRequest;
  }

  get isDeletable(): boolean {
    return true;
  }

  get descriptionIsEditable(): boolean {
    return true;
  }

  get renderedDescription(): Node {
    return <>{this.description}</>;
  }
}